Add `endpoint` option to `Builder`
Currently the Builder's initializer accepts a host param
which assumes https. I like to use minio for local
development and it isn't ideal to run it with https.

The normal Aws::S3::Client gem accepts an :endpoint
option to by-pass this issue.

Here is a section of the docs that explains this option:

```
:endpoint (String, URI::HTTPS, URI::HTTP) — Normally you should not configure the :endpoint option directly. This is normally constructed from the :region option. Configuring :endpoint is normally reserved for connecting to test or custom endpoints. The endpoint should be a URI formatted like:
'http://example.com'
'https://example.com'
'http://example.com:123'
```
https://docs.aws.amazon.com/sdk-for-ruby/v3/api/Aws/S3/Client.html

In this commit I add the extra option into the builder class while
maintaining the current functionality for the host option.