Refactor uri logic out of the Builder
The builder class is pretty large and it has to do a lot of extra edges
cases around finding the correct host and endpoint. In this commit I
create a FasterS3Url::Uri class and move all uri specific logic to that
class.

The builder class caches it and all of it's attributes are cached, so it
shouldn't have a negative impact on performance.